refactor(providers): Refactor OpenAI provider API delegates and settings structure
- Standardize HTTP client configuration across all OpenAI API delegates.
- Introduce dedicated API delegates for completions and responses functionality.
- Simplify settings by moving from `httpVersionConfig` to `httpVersion`.
- Apply minor UI styling adjustments in the desktop application.
